pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantbot"
version = "0.1.0"
description = "Multi-timeframe quant trading bot skeleton with news + orderbook hooks"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "httpx>=0.27",
  "websockets>=12.0",
  "APScheduler>=3.10",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "feedparser>=6.0.11",
  "PyJWT>=2.8",
  "rich>=13.7",
]

[project.scripts]
quantbot = "quantbot.main:main"
